Fix hipe memory allocation problems on darwin/amd64.
HiPE on amd64 needs to be able to allocate memory in the 0x20000000+
range so that compiled code will be addressable with 32bit pointers
(when sign extended). By default, when building 64bit executables on OS X, PAGEZERO
occupies the first 4GB of virtual memory. The PAGEZERO segment cannot
be mmap(ed) or vm_allocate(d) into.

To avoid this problem we use the -pagezero_size linker flag to reduce
the size of PAGEZERO and free up the 0x20000000+ space.