Fix re:replace/4 to handle unicode charlist RE argument
A bug in re:replace/4 causes a badarg exception to be thrown when
the unicode option is selected and the RE argument is a charlist
containing non-ascii codepoints.

The problem occurs because re:replace/4 flattens it's Subject
argument to a binary() and then calls re:run/3 with this new
flattened Subject. This causes an exception when the RE argument
is a unicode charlist(), because the rules for re:run/3 specify
that Subject must be a valid charlist() when RE is a charlist().
A binary() is not a valid charlist().

This patch does not address the underlying problem which is that
re:run/3 requires a charlist() when unicode is being used. A better
solution would be to change re:run/3 to accept any chardata()
(ie charlist() or unicode-binary()).

NOTE: There is also a similar bug in re:split/3 which also changes
it's Subject to a binary() before calling re:run/3.  However, this
bug will not show up under testing at the moment.  It's hidden by
a different bug somewhere in re:compile, where the unicode flag
is ignored.