First experiment with end-to-end file command tagging with DTrace.
The experiment is to somehow tag a process that will later
execute the file:rename/2 function.  (The API could have to change
if the tag were explicitly added to rename()'s arguments somehow.)
The rename() function is executed via the code server, so we have
to deal with a proxy process, a.k.a. the file server (registered
under the name 'file_server_2').

I'm not exactly happy with this choice of tagging, namely using
`put(dtrace_utag, Tag::list())`.  But this is an experiment, so
we'll see how it goes.  I can't imagine changing the API for
all file.erl functions in order pass the tag explicitly.

Demo:

In a root login shell:

    # dtrace -xevaltime=exec -Z -n 'erlang_vm*:::file_drv_entry {printf("%d %d %s | %d | %s %s , %d %d", arg0, arg1, arg2 == NULL ? "" : copyinstr(arg2), g3, arg4 == NULL ? "" : copyinstr(arg4), arg5 == NULL ? "" : copyinstr(arg5), arg6, arg7)} erlang_vm*:::file_drv_int* {printf("%d %d %d | %d", arg0, arg1, arg2, arg3);} erlang_vm*:::file_drv_return {printf("%d %d %s | %d | %d %d %d", arg0, arg1, arg2 == NULL ? "" : copyinstr(arg2), arg3, arg4, arg5, arg6 ) ; } erlang_vm*:::xx {printf("%s", arg1 == NULL ? "-null-" : copyinstr(arg1));}'

In an Erlang shell:

    1> put(dtrace_utag, "GGOOOAAALL!!!!!").
    undefined
    2> file:rename("asdf", "zxcv").
    {error,enoent}

Example output:

      0 132464       file_output:file_drv_entry 0 517 GGOOOAAALL!!!!! | 12 | asdf zxcv , 0 0
      1 132470 invoke_rename:file_drv_int_entry 0 517 5004 | 12
      1 132497 invoke_rename:file_drv_int_return 0 517 5004 | 12
      1 132518 file_async_ready:file_drv_return 0 517 GGOOOAAALL!!!!! | 12 | 0 2 5004

Also, add DTrace branch explanation to README.md