Edits to wrap DTrace usage in DTRACEn() and DTRACE_ENABLED() macros
This was a difficult decision to make: choose the straightforward
macro style that Dustin and Michal used, or use the style macros
that I first adopted for the efile_drv.c driver.

Either choice would require a fair amount of edits.  However, what
finally tipped my decision was Michal's commit aabf21a37e,
"Disabling/enabling DTrace support in compile time".  Much of that
commit was adding #ifdef/#endif around various DTrace macros.
Almost all of the edits in that commit could be eliminated by using
the DTRACEn() and DTRACE_ENABLED() macros.  So, on the basis of
"cleaner" code with fewer #ifdef/#endifs, I made the choice.

Also in this commit: various small edits to fix OS X gcc compiler
warnings and to cleanly compile with and without "--enable-dtrace"
configured.