Relocate the erlang_dtrace.h (generated) file.
Gadz, wrestling with autoconf and make isn't fun.  But I believe
that this patch successfully moves the generated-by-`dtrace`
file `erlang_dtrace.h` to a platform-specific directory.  The
new location should be much less odious to the OTP team.

Also, this patch is starting to make inroads to making DTrace-
via-SystemTap compilation on Linux platforms possible.