Add tentative support for sequential tracing when sending a message.
This implementation uses a struct typedef.  I don't think I like it:
there's lots of cut-and-paste.  I could avoid the cut-and-paste by
putting the definition in a single central place, but then using it
with D scripts becomes complicated because you need to specify using
the CPP and then specifying the proper path to find the central
header file.  And I don't believe that I like that CPP complication
any better than I like the cut-and-paste ... so I'll probably yank
this stuff out and replace it with simple scalars in the next commit.