Small cleanups per GitHub pull request #1, part 4
We cannot easily use the DTRACE_TERM_BUF_SIZE macro to declare the
size of `erts_this_node_sysname` because that macro would also
appear in `erl_node_tables.h`, which is (indirectly) included by
just about every single source file, which would then add a
dependency on the `dtrace-wrapper.h` macro.  So, we'll use a
constant and a couple of comment warnings.