Move 2 USDT probes from dtrace_user.d -> erlang_dtrace.d
This is a move that pollutes the Erlang VM with functions that are
only used by the NIF shared library code in lib/dtrace/c_src/dtrace.c.
Both probes are for use by Erlang code to fire DTrace probes, e.g.,
`dtrace:p("Hello, world!")`.

The reason for this not-so-wonderful move is that the current
code (before this commit) does not work correctly on all supported
platforms.  Here's the list of platforms and their before-this-commit
status:

	CentOS 5/SystemTap 1.3		Doesn't work
	CentOS 6/SystemTap 1.2		Works fine
	CentOS 6/SystemTap 1.4		Works fine
	FreeBSD 9.0RC2			Works fine
	OS X				Works fine
	Solaris 10			Doesn't work
	Solaris 11			Doesn't work
	Open Solaris OpenIndiana 151a	Doesn't work
	Illumos/SmartOS			Doesn't work

I haven't figured out why the shared library probes don't work
with SystemTap 1.3 but do work with 1.2 and 1.4.

The Solaris issue is discussed in this `dtrace-discuss` mailing
list thread:
http://www.listbox.com/member/archive/184261/2011/12/sort/time_rev/page/1/entry/3:5/20111201185820:57584502-1C78-11E1-96DD-B85D9B43C33E/

This commit moves the DTRACE_ENABLED() and DTRACE() macro invocations
into statically linked code inside the virtual machine.  When the
probes are fired via the NIF at their new location, the probes work
correctly.